+++ embark/workers/models.py
@@ -1,10 +1,8 @@
-import re
 import ipaddress
 import socket
 import paramiko
 
 from django.db import models
-from django.forms import ModelForm
 from django.core.exceptions import ValidationError
 from django.utils.translation import gettext_lazy as _
 
@@ -19,19 +17,6 @@ class Configuration(models.Model):
     ssh_password = models.CharField(max_length=150)
     ip_range = models.CharField(max_length=150)
     created_at = models.DateTimeField(auto_now_add=True)
-
-
-class ConfigurationForm(ModelForm):
-    class Meta:
-        model = Configuration
-        fields = ['name', 'ssh_user', 'ssh_password', 'ip_range']
-
-    def clean_ip_range(self):
-        ip_range = self.cleaned_data.get('ip_range')
-        ip_range_regex = r"^(\d{1,3}\.){3}\d{1,3}/\d{1,2}$"
-        if not re.match(ip_range_regex, ip_range):
-            raise ValidationError("Invalid IP range format. Use CIDR notation")
-        return ip_range
 
 
 def default_deb_list():
